requirements/requirements.txt
tortoise-orm[aiosqlite]>=0.20.0
python-dotenv>=1.0.0
aiogram>=3.4.0
dnspython>=2.7.0
fastapi>=0.115.12
uvicorn>=0.34.0
